pyproject.toml
[project]
name = "morphik-code-exec"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "isort>=7.0.0",
    "morphik>=1.1.2",
    "openai>=2.15.0",
    "python-dotenv>=1.2.1",
]
